third_party/blink/web_tests/fast/hidpi/image-srcset-same-alternative-for-both-attributes.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }
    addEventListener("load", function() {
        if (internals)
            shouldBeFalse('internals.isPreloaded("resources/green-400-px-square.png")');
        shouldBeTrue('document.getElementById("foo").clientWidth==100');
    }, false);
</script>
<div>This test passes if this img tag below is a blue square when the scale factor is 1. It ensures that when both source attributes has the same input alternative always the one from srcset is chosen</div>
<img id="foo" src="resources/green-400-px-square.png" srcset="resources/blue-100-px-square.png 1x">
